refactor: split utils into submodules feat: retry when hitting flood limit enhance: add error handler
