+++ nazurin/bot.py
@@ -1,14 +1,20 @@
-from typing import List, Union
+from html import escape
+from mimetypes import guess_type
+from typing import List
 
 from aiogram import Bot, Dispatcher, executor
-from aiogram.dispatcher.filters import BoundFilter
 from aiogram.dispatcher.handler import CancelHandler
 from aiogram.dispatcher.middlewares import BaseMiddleware
-from aiogram.types import AllowedUpdates, Message
+from aiogram.types import (AllowedUpdates, ChatActions, InputFile,
+                           InputMediaPhoto, Message)
+from aiogram.utils.exceptions import BadRequest
 from aiogram.utils.executor import start_webhook
 
 from nazurin import config
-from nazurin.utils import getUrlsFromMessage, logger
+from nazurin.utils import logger
+from nazurin.utils.decorators import chat_action, retry_after
+from nazurin.utils.filters import URLFilter
+from nazurin.utils.helpers import handleBadRequest
 
 _HANDLER_ATTRS = '__nazurin.handler'
 
@@ -18,6 +24,54 @@ def handler(*args, **kwargs):
         return callback
 
     return decorator
+
+@chat_action(ChatActions.UPLOAD_PHOTO)
+@retry_after
+async def sendPhotos(message: Message, imgs: List['Image'], details=None):
+    if details is None:
+        details = dict()
+    media = list()
+    if len(imgs) > 10:
+        # TODO
+        imgs = imgs[:10]
+        await message.reply('Notice: Too many pages, sending only 10 of them')
+
+    caption = str()
+    for key, value in details.items():
+        caption += str(key) + ': ' + str(value) + '\n'
+    if len(caption) > 1024:
+        caption = caption[:1024]
+        await message.reply('Notice: Caption too long, trimmed')
+    caption = escape(caption, quote=False)
+
+    for img in imgs:
+        filetype = str(guess_type(img.url)[0])
+        if filetype.startswith('image'):
+            media.append(InputMediaPhoto(img.display_url))
+        else:
+            await message.reply('File is not image, try download option.')
+            return
+    media[0].caption = caption
+    try:
+        await message.reply_media_group(media)
+    except BadRequest as error:
+        await handleBadRequest(message, error)
+
+@retry_after
+async def sendDocument(message: Message, img: 'Image', chat_id, message_id):
+    await message.bot.send_document(chat_id,
+                                    InputFile(img.path),
+                                    reply_to_message_id=message_id)
+
+@chat_action(ChatActions.UPLOAD_DOCUMENT)
+async def sendDocuments(message: Message, imgs: List['Image'], chat_id=None):
+    message_id = message.message_id
+    if not chat_id:
+        chat_id = message.chat.id
+    else:
+        message_id = None  # Sending to channel, no message to reply
+    for img in imgs:
+        await sendDocument(message, img, chat_id, message_id)
 
 class AuthMiddleware(BaseMiddleware):
     async def on_process_message(self, message: Message, data: dict):
@@ -29,18 +83,12 @@ class AuthMiddleware(BaseMiddleware):
             return
         raise CancelHandler()
 
-class URLFilter(BoundFilter):
-    key = 'url'
-
-    async def check(self, message: Message) -> Union[List[str], bool]:
-        urls = getUrlsFromMessage(message)
-        if urls:
-            return {'urls': urls}
-        return False
+class NazurinBot(Bot):
+    send_message = retry_after(Bot.send_message)
 
 class Nazurin(Dispatcher):
     def __init__(self):
-        bot = Bot(token=config.TOKEN)
+        bot = NazurinBot(token=config.TOKEN)
         super().__init__(bot)
         self.middleware.setup(AuthMiddleware())
         self.filters_factory.bind(URLFilter,
@@ -56,9 +104,10 @@ class Nazurin(Dispatcher):
                                    allowed_updates=AllowedUpdates.MESSAGE)
 
     async def on_shutdown(self, dp):
-        await self.bot.delete_webhook()
+        await self.bot.delete_webhook()  # TODO
 
     def start(self):
+        # TODO
         if config.ENV == 'production':
             start_webhook(self,
                           webhook_path='/' + config.TOKEN,
