+++ nazurin/sites/Gelbooru/api.py
@@ -3,7 +3,8 @@ from typing import List
 import requests
 
 from nazurin.models import Image
-from nazurin.utils import NazurinError, downloadImages
+from nazurin.utils import downloadImages
+from nazurin.utils.exceptions import NazurinError
 
 
 class Gelbooru(object):
