+++ nazurin/sites/Moebooru/commands.py
@@ -2,7 +2,7 @@ from aiogram.dispatcher import filters
 from aiogram.types import Message
 
 from nazurin import bot
-from nazurin.utils import NazurinError, sendDocuments, sendPhotos
+from nazurin.utils.exceptions import NazurinError
 
 from .api import Moebooru
 
@@ -16,13 +16,11 @@ async def yandere_view(message: Message, regexp_command):
             await message.reply('Invalid post id!')
             return
         imgs, details = moebooru.site('yande.re').view(post_id)
-        await sendPhotos(message, imgs, details)
+        await bot.sendPhotos(message, imgs, details)
     except (IndexError, ValueError):
         await message.reply('Usage: /yandere <post_id>')
     except NazurinError as error:
         await message.reply(error.msg)
-    # except BadRequest as error:
-    #     handleBadRequest(message, error)
 
 @bot.handler(
     filters.RegexpCommandsFilter(regexp_commands=[r'/yandere_download (\S+)']))
@@ -33,7 +31,7 @@ async def yandere_download(message: Message, regexp_command):
             await message.reply('Invalid post id!')
             return
         imgs = await moebooru.site('yande.re').download(post_id)
-        await sendDocuments(message, imgs)
+        await bot.sendDocuments(message, imgs)
     except (IndexError, ValueError):
         await message.reply('Usage: /yandere_download <post_id>')
     except NazurinError as error:
@@ -48,13 +46,11 @@ async def konachan_view(message: Message, regexp_command):
             await message.reply('Invalid post id!')
             return
         imgs, details = moebooru.site('konachan.com').view(post_id)
-        await sendPhotos(message, imgs, details)
+        await bot.sendPhotos(message, imgs, details)
     except (IndexError, ValueError):
         await message.reply('Usage: /konachan <post_id>')
     except NazurinError as error:
         await message.reply(error.msg)
-    # except BadRequest as error:
-    #     handleBadRequest(message, error)
 
 @bot.handler(
     filters.RegexpCommandsFilter(regexp_commands=[r'/konachan_download (\S+)'])
@@ -66,7 +62,7 @@ async def konachan_download(message: Message, regexp_command):
             await message.reply('Invalid post id!')
             return
         imgs = await moebooru.site('konachan.com').download(post_id)
-        await sendDocuments(message, imgs)
+        await bot.sendDocuments(message, imgs)
     except (IndexError, ValueError):
         await message.reply('Usage: /konachan_download <post_id>')
     except NazurinError as error:
