+++ nazurin/sites/Zerochan/commands.py
@@ -2,7 +2,7 @@ from aiogram.dispatcher import filters
 from aiogram.types import Message
 
 from nazurin import bot
-from nazurin.utils import NazurinError, sendDocuments, sendPhotos
+from nazurin.utils.exceptions import NazurinError
 
 from .api import Zerochan
 
@@ -17,13 +17,11 @@ async def zerochan_view(message: Message, regexp_command):
             await message.reply('Invalid post id!')
             return
         imgs, details = api.view(post_id)
-        await sendPhotos(message, imgs, details)
+        await bot.sendPhotos(message, imgs, details)
     except (IndexError, ValueError):
         await message.reply('Usage: /zerochan <post_id>')
     except NazurinError as error:
         await message.reply(error.msg)
-    # except BadRequest as error:
-    #     handleBadRequest(message, error)
 
 @bot.handler(
     filters.RegexpCommandsFilter(regexp_commands=[r'/zerochan_download (\S+)'])
@@ -35,7 +33,7 @@ async def zerochan_download(message: Message, regexp_command):
             await message.reply('Invalid post id!')
             return
         imgs = await api.download(post_id)
-        await sendDocuments(message, imgs)
+        await bot.sendDocuments(message, imgs)
     except (IndexError, ValueError):
         await message.reply('Usage: /zerochan_download <post_id>')
     except NazurinError as error:
