+++ nazurin/utils/exceptions.py
@@ -1,0 +1,9 @@
+class NazurinError(Exception):
+    def __init__(self, msg):
+        """Initialize with error message."""
+        super().__init__(msg)
+        self.msg = str(msg)
+
+    def __str__(self):
+        """Returns the string representation of this exception."""
+        return self.msg
